refactor(modifiers): isolate authority math
